requirements.txt
numpy==1.26.4
pandas==2.1.4
openpyxl==3.1.2
blabel==0.1.5
reportlab>=4.0.0
Pillow==10.0.0
python-barcode==0.14.0
